NavBar moved to container folder
